stock_api_backend/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.4
djangorestframework==3.15.2
dnspython==2.7.0
idna==3.10
pymongo==4.10.1
python-decouple==3.8
requests==2.32.3
sqlparse==0.5.2
urllib3==2.2.3
